--- a/shared/templates/create_permission.py
+++ b/shared/templates/create_permission.py
@@ -5,13 +5,12 @@
 #        generate template-based checks for file permissions/ownership
 
 
-import sys
 import re
 
 from template_common import FilesGenerator, UnknownTargetError
 
-class PermissionGenerator(FilesGenerator):
 
+class PermissionGenerator(FilesGenerator):
     def generate(self, target, path_info):
         # the csv file contains lines that match the following layout:
         #    directory,file_name,uid,gid,mode
@@ -22,7 +21,7 @@ class PermissionGenerator(FilesGenerator):
         # path_id maps to FILEID in the template
         if file_name == '[NULL]':
             path_id = re.sub('[-\./]', '_', dir_path)
-        elif re.match( r'\^.*\$', file_name, 0):
+        elif re.match(r'\^.*\$', file_name, 0):
             path_id = re.sub('[-\./]', '_', dir_path) + '_' + re.sub('[-\\\./^$*(){}|]',
                                                                      '_', file_name)
             # cleaning trailing end multiple underscores, make sure id is lowercase
@@ -42,7 +41,7 @@ class PermissionGenerator(FilesGenerator):
             full_path = dir_path + '/' + file_name
 
         if target == "bash":
-            if not re.match( r'\^.*\$', file_name, 0):
+            if not re.match(r'\^.*\$', file_name, 0):
                 self.file_from_template(
                     "./template_BASH_permissions",
                     {
@@ -52,7 +51,7 @@ class PermissionGenerator(FilesGenerator):
                     "./bash/file_permissions{0}.sh", path_id
                 )
             else:
-                file_name = re.sub('^\^','', file_name)
+                file_name = re.sub('^\^', '', file_name)
                 self.file_from_template(
                     "./template_BASH_regex_permissions",
                     {
@@ -63,7 +62,7 @@ class PermissionGenerator(FilesGenerator):
                     "./bash/file_permissions{0}.sh", path_id
                 )
 
-        elif target == "ansible" and not re.match( r'\^.*\$', file_name, 0):
+        elif target == "ansible" and not re.match(r'\^.*\$', file_name, 0):
             self.file_from_template(
                 "./template_ANSIBLE_permissions",
                 {
@@ -118,8 +117,5 @@ class PermissionGenerator(FilesGenerator):
 
     def csv_format(self):
         return("CSV should contains lines of the format: "
-              "directory path,file name,owner uid (numeric),group "
-              "owner gid (numeric),mode")
-
-if __name__ == "__main__":
-    PermissionGenerator().main()
+               "directory path,file name,owner uid (numeric),group "
+               "owner gid (numeric),mode")