--- a/shared/templates/create_services_enabled.py
+++ b/shared/templates/create_services_enabled.py
@@ -96,7 +96,3 @@ class ServiceEnabledGenerator(FilesGenerator):
     def csv_format(self):
         return("CSV should contains lines of the format: " +
                "servicename,packagename")
-
-
-if __name__ == "__main__":
-    ServiceEnabledGenerator().main()