--- a/shared/utils/generate-from-templates.py
+++ b/shared/utils/generate-from-templates.py
@@ -3,6 +3,7 @@
 import os
 import sys
 import argparse
+import inspect
 
 templates_dir = os.path.join(os.path.dirname(__file__), "..", "templates")
 sys.path.append(templates_dir)
@@ -142,6 +143,9 @@ class Builder(object):
         assert(action in [ActionType.INPUT, ActionType.OUTPUT])
 
         list_ = []
+        if action == ActionType.INPUT:
+            # add template_common source file
+            list_.append(os.path.abspath(inspect.getsourcefile(ActionType)))
 
         for oval in self.supported_ovals:
             self._set_current_oval(oval)