We run template generators separately as scripts and then parse what they output on stdout. This hides errors and warnings and is much slower than just using the classes and functions directly. I am not sure what the reason for running a script that runs more scripts was but I made a quick refactor to just use the generators as classes. This results in less code and it's much faster.

~~This PR is not finished, I want to clean it all up and write it more elegantly. But first I want consensus from the community that this is the way to go. Maybe there is a reason why we run all the scripts separately.~~

Configure time:
```
old code: 51s
new code: 26s
```

Build time:
```
old code: 3m 23s
new code: 2m 54s
```

(I am working on battery, the numbers plugged in are going to be better, just included this for a comparison)