@mpreisler 2017-06-26
plugged in the numbers are:
```
configure: 16s
build: 2m 3s
```

I would love some feedback from the community. This seems like the right way to go, it already uncovered a bunch of suspect things. For example we build packages_removed in shared csv but these are built again for the products where they belong. The packages_removed.csv in shared templates can't possibly generated OVAL because the OVAL is different for RHEL and Debian.