@jan-cerny 2017-06-27
This looks very interesting, The speedup is great.

I'm not aware of any reason why the scripts are run separately.