@mpreisler 2017-06-27
This breaks the build in a bizarre way. It's building the OVAL unlinked target 3 times when these changes are applied. If anybody can tell me why I would appreciate it.