@mpreisler 2017-06-27
Will wait for #2116 before I debug the weird build issue, it will make the build graph way simpler.