@mpreisler 2017-06-27
What a mystery:
- same targets on master and this branch
- same generated dirs - $CMAKE_CURRENT_BUILD_DIR/checks and remediations are the same, exactly the same
- same inputs except this branch is missing the python script deps, that shouldn't matter for this issue though
- exactly the same outputs

wtf is going on...
```