@mpreisler 2017-06-27
OK, I figured it out, the outputs weren't the same. We included outputs for things that were missing templates so they were never actually written out. That's why it kept rebuilding oval-unlinked.xml