@mpreisler 2017-06-27
There is some new issue with dep rescan. OVAL gets rebuilt even if none of the files changed... Need to figure out what's going on and fix it before we merge.