@mpreisler 2017-06-27
After this is merged by far the biggest time wasters will be OVAL validations:
```
262.925 RHEL/5/validation-ssg-rhel5-oval.xml
249.489 RHEL/6/validation-ssg-rhel6-oval.xml
235.243 RHEL/7/validation-ssg-rhel7-oval.xml
144.528 OpenStack/RHEL-OSP/7/validation-ssg-rhel-osp7-oval.xml
123.681 Fedora/validation-ssg-fedora-oval.xml
54.595 Ubuntu/14.04/validation-ssg-ubuntu1404-oval.xml
54.04 Debian/8/validation-ssg-debian8-oval.xml
51.334 Ubuntu/16.04/validation-ssg-ubuntu1604-oval.xml
47.8035 ssg-rhel6-ds.xml
41.4005 ssg-rhel7-ds.xml
```

time in seconds. OVAL schematron takes forever :-(