@mpreisler 2017-06-28
Rebased & fixed conflicts