@dahaic 2017-06-28
So the schematron rules with the O(n^3) and O(n^2), being official is something we should be running. But at the same time there is no reason to stick to it as a PR gating mechanism - what about making our own, simplified, schematron rules for PR gating, which will catch most common issues that happen during SSG development (that will be some small subset of all possible errors, due to the way our build system works).

And leaving official check to be run every day. And the very small subset of errors not yet catched by our gating can be fixed by us. (Together with gating schematron update).