@mpreisler 2017-06-28
@dahaic  I like the idea. I would even go as far as to avoid schematron for PR gating and only do it nightly. XSD will catch 99% of all issues, the schematron is for very special and rare issues.

I will implement a PR that makes schematron checking an option via cmake and default it to disabled. We can discuss this further there.